+++ modulo3/projeto-astromatch/astromatch/src/components/Limpar.js
@@ -23,12 +23,10 @@ export default function Limpar(props) {
 //Estilização Botão
 const Botao = Styled.button`
 padding: 4px 11px;
-
 border: 1px solid #ffcc00;
 -moz-border-radius: 8px;
 -webkit-border-radius: 8px;
 border-radius: 8px;
-
 background-color: #861405 !important;
 background: -webkit-gradient(linear, left top, left bottom, color-stop(0%,#FEBEB6), color-stop(100%,#861405)) !important;
 background: -webkit-linear-gradient(top, #FEBEB6, ##861405) !important;
@@ -39,10 +37,12 @@ font-family:"Segoe UI";
 font-weight: bold;
 color: #fff;
 text-shadow: 0 1px 0 #666;
-
 box-shadow: inset 0 1px 1px #fff, 0 2px 3px #666;
 cursor:pointer;
-position: relative;
+position: absolute;
+bottom: 2em;
+left:11em;
+
 
 :hover {
 box-shadow: inset 0 2px 5px #fff, 0 2px 3px #666;
@@ -51,6 +51,5 @@ background: -webkit-gradient(linear, left top, left bottom, from(#FEBEB6), to(##
 
 :active{
 box-shadow: inset 0 2px 5px #fff, 0 1px 3px #666 !important;
-top:2px
 }
 `
